sched: ignore task_h_load for CPU_NEWLY_IDLE
A load balancer calculates imbalance factor for particular sched
domain and tries to steal up the prescribed amount of weighted load.
However, a small imbalance factor would sometimes prevent us from
stealing any tasks at all. When a CPU is newly idle, it should
steal first task which meets the migration criteria.

There is a slight improvement when it comes to frame drops
(in my case drops per/two seconds). Basically a test case is
left finger swipe on the display (21 times, duration is 2
seconds + 1 second sleep between iterations):

0   Framedrops:  7    5
1   Framedrops:  5    3
2   Framedrops:  8    5
3   Framedrops:  4    5
4   Framedrops:  3    3
5   Framedrops:  6    4
6   Framedrops:  3    2
7   Framedrops:  3    4
8   Framedrops:  5    3
9   Framedrops:  3    3
10  Framedrops:  7    4
11  Framedrops:  3    4
12  Framedrops:  3    3
13  Framedrops:  3    3
14  Framedrops:  3    5
15  Framedrops:  7    3
16  Framedrops:  5    3
17  Framedrops:  3    2
18  Framedrops:  5    3
19  Framedrops:  4    3
20  Framedrops:  3    2

max is 8 vs 5; min is 3 vs 2.
As for applied load, it is not significant and is "light".

Signed-off-by: Uladzislau 2 Rezki <uladzislau2.rezki@sonymobile.com>
Signed-off-by: Artem Labazov <123321artyom@gmail.com>
Signed-off-by: celtare21 <celtare21@gmail.com>