cpuidle: lpm-levels: refactor cpu count casts
Qualcomm, the true geniuses they are, decided to initially cast this
variable to an unsigned long, and then once again to an unsigned int for
each usage case. Not only does this allocate the bit width of an
unsigned long for an integer value, we also risk an alignment fault from
converting a long to an integer. Let's use the proper type for this,
uintprt_t.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>